refactor: move topics system from pkg/cobrax to cmd/dodot/internal
Move the topics help system from pkg/cobrax/topics to cmd/dodot/internal/topics since it's CLI-specific functionality rather than a reusable package.

Changes:
- Moved all files from pkg/cobrax/topics/ to cmd/dodot/internal/topics/
- Updated import in cmd/dodot/root.go to use new location
- Removed empty pkg/cobrax directory
- All functionality preserved (help topics, paging, rendering)
- All tests continue to pass (856 tests successful)
