+++ cmd/dodot/root.go
@@ -18,8 +18,8 @@ import (
 	"github.com/arthur-debert/dodot/cmd/dodot/commands/snippet"
 	"github.com/arthur-debert/dodot/cmd/dodot/commands/status"
 	"github.com/arthur-debert/dodot/cmd/dodot/commands/topics"
+	topicspkg "github.com/arthur-debert/dodot/cmd/dodot/internal/topics"
 	"github.com/arthur-debert/dodot/internal/version"
-	topicspkg "github.com/arthur-debert/dodot/pkg/cobrax/topics"
 	"github.com/arthur-debert/dodot/pkg/commands"
 	offpkg "github.com/arthur-debert/dodot/pkg/commands/off"
 	onpkg "github.com/arthur-debert/dodot/pkg/commands/on"
